api/pyproject.toml
[project]
name = "motoko-chat"
version = "1.0.0"
description = "Motoko web chat interface"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "sse-starlette>=2.0.0",
    "pyjwt>=2.8.0",
    "bcrypt>=4.1.0",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0.0",
    "motoko @ file://..",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
include = ["*.py", "static/**/*"]
